COMP: Fix compilation error with ITK 4.8.1 on Mac OS X

+++ src/Components/Optimizers/AdaptiveStochasticGradientDescent/elxAdaptiveStochasticGradientDescent.hxx
@@ -170,8 +170,8 @@ AdaptiveStochasticGradientDescent< TElastix >
     /** Set the maximum step length: the maximum displacement of a voxel in mm.
      * Compute default value: mean in-plane spacing of fixed and moving image.
      */
-    const unsigned int fixdim = vnl_math_min( this->GetElastix()->FixedDimension, (unsigned int) 2 );
-    const unsigned int movdim = vnl_math_min( this->GetElastix()->MovingDimension, (unsigned int) 2 );
+    const unsigned int fixdim = vnl_math_min( (unsigned int) this->GetElastix()->FixedDimension, (unsigned int) 2 );
+    const unsigned int movdim = vnl_math_min( (unsigned int) this->GetElastix()->MovingDimension, (unsigned int) 2 );
     double             sum    = 0.0;
     for( unsigned int d = 0; d < fixdim; ++d )
     {
